+++ packages/forms/signals/src/api/rules/index.ts
@@ -10,5 +10,6 @@ export * from './aggregate_metadata';
 export * from './disabled';
 export * from './hidden';
 export * from './readonly';
-export * from './validate';
+export * from './validation/validate';
 export * from './validate_tree';
+export * from './validation';
