+++ packages/forms/signals/src/api/rules/validation/index.ts
@@ -14,3 +14,5 @@ export * from './min_length';
 export * from './pattern';
 export * from './required';
 export * from './standard_schema';
+export * from './validate_async';
+export * from './validate_http';
