+++ packages/forms/signals/src/api/rules/validation/max_length.ts
@@ -7,10 +7,11 @@
  */
 
 import {computed} from '@angular/core';
-import {aggregateMetadata, validate} from '../logic';
-import {metadata, MIN_LENGTH} from '../metadata';
-import {SchemaPath, LogicFn, PathKind, SchemaPathRules} from '../types';
-import {minLengthError} from '../validation_errors';
+import {aggregateMetadata} from '../aggregate_metadata';
+import {validate} from './validate';
+import {metadata, MAX_LENGTH} from '../metadata';
+import {SchemaPath, SchemaPathRules, LogicFn, PathKind} from '../../types';
+import {maxLengthError} from './validation_errors';
 import {
   BaseValidatorConfig,
   getLengthOrSize,
@@ -20,15 +21,15 @@ import {
 } from './util';
 
 /**
- * Binds a validator to the given path that requires the length of the value to be greater than or
- * equal to the given `minLength`.
+ * Binds a validator to the given path that requires the length of the value to be less than or
+ * equal to the given `maxLength`.
  * This function can only be called on string or array paths.
- * In addition to binding a validator, this function adds `MIN_LENGTH` property to the field.
+ * In addition to binding a validator, this function adds `MAX_LENGTH` property to the field.
  *
  * @param path Path of the field to validate
- * @param minLength The minimum length, or a LogicFn that returns the minimum length.
+ * @param maxLength The maximum length, or a LogicFn that returns the maximum length.
  * @param config Optional, allows providing any of the following options:
- *  - `error`: Custom validation error(s) to be used instead of the default `ValidationError.minLength(minLength)`
+ *  - `error`: Custom validation error(s) to be used instead of the default `ValidationError.maxLength(maxLength)`
  *    or a function that receives the `FieldContext` and returns custom validation error(s).
  * @template TValue The type of value stored in the field the logic is bound to.
  * @template TPathKind The kind of path the logic is bound to (a root path, child path, or item of an array)
@@ -36,31 +37,31 @@ import {
  * @category validation
  * @experimental 21.0.0
  */
-export function minLength<
+export function maxLength<
   TValue extends ValueWithLengthOrSize,
   TPathKind extends PathKind = PathKind.Root,
 >(
   path: SchemaPath<TValue, SchemaPathRules.Supported, TPathKind>,
-  minLength: number | LogicFn<TValue, number | undefined, TPathKind>,
+  maxLength: number | LogicFn<TValue, number | undefined, TPathKind>,
   config?: BaseValidatorConfig<TValue, TPathKind>,
 ) {
-  const MIN_LENGTH_MEMO = metadata(path, (ctx) =>
-    computed(() => (typeof minLength === 'number' ? minLength : minLength(ctx))),
+  const MAX_LENGTH_MEMO = metadata(path, (ctx) =>
+    computed(() => (typeof maxLength === 'number' ? maxLength : maxLength(ctx))),
   );
-  aggregateMetadata(path, MIN_LENGTH, ({state}) => state.metadata(MIN_LENGTH_MEMO)!());
+  aggregateMetadata(path, MAX_LENGTH, ({state}) => state.metadata(MAX_LENGTH_MEMO)!());
   validate(path, (ctx) => {
     if (isEmpty(ctx.value())) {
       return undefined;
     }
-    const minLength = ctx.state.metadata(MIN_LENGTH_MEMO)!();
-    if (minLength === undefined) {
+    const maxLength = ctx.state.metadata(MAX_LENGTH_MEMO)!();
+    if (maxLength === undefined) {
       return undefined;
     }
-    if (getLengthOrSize(ctx.value()) < minLength) {
+    if (getLengthOrSize(ctx.value()) > maxLength) {
       if (config?.error) {
         return getOption(config.error, ctx);
       } else {
-        return minLengthError(minLength, {message: getOption(config?.message, ctx)});
+        return maxLengthError(maxLength, {message: getOption(config?.message, ctx)});
       }
     }
     return undefined;
