+++ packages/forms/signals/test/node/api/validators/min_length.spec.ts
@@ -9,7 +9,7 @@
 import {Injector, signal} from '@angular/core';
 import {TestBed} from '@angular/core/testing';
 import {MIN_LENGTH, form, minLength} from '../../../../public_api';
-import {customError, minLengthError} from '../../../../src/api/validation_errors';
+import {customError, minLengthError} from '../../../../src/api/rules/validation/validation_errors';
 
 describe('minLength validator', () => {
   it('returns minLength error when the length is smaller for strings', () => {
