+++ packages/forms/signals/test/node/recursive_logic.spec.ts
@@ -9,7 +9,7 @@
 import {computed, Injector, signal, type Signal} from '@angular/core';
 import {TestBed} from '@angular/core/testing';
 import {customError} from '../../public_api';
-import {disabled, validate} from '../../src/api/logic';
+import {disabled, validate} from '../../src/api/rules';
 import {applyEach, applyWhen, applyWhenValue, form, schema} from '../../src/api/structure';
 import type {FieldTree, Schema} from '../../src/api/types';
